feat(content-system): judge the selected translation entry and trim the doc surfaces
Six surfaces the translation work edited ended above their size budget, and the reduction-order rule was restated in full on eight surfaces while docs/pipeline-steps.md owns it. Each of Binding, Layout/Type, Diagnostics and Mutation moves its symbol index into a docs/symbols.md file, Diagnostics moves its per-violation reporting rules into docs/violation-rules.md, the listener guide splits its RenderedElement API into docs/listener-api.md, and the module README moves its reference index into docs/README.md. The non-owner surfaces keep the one local consequence each exists to state plus a link to the owner.

The PropertyType class docblock said the translatable flag is ignored on a non-primitive type; the validator rejects it and admits() branches on it first, so the sentence now states that. PropertyTypeConformanceValidator drops an is_array term unreachable after admits() accepted a translatable value. History words in four surfaces and one test docblock are rephrased in present tense.

Language reduction judged only the outer shape of a translatable value, so a map whose selected entry is a nested map or a number was served verbatim instead of failing like a bare string does. The selected entry must be a string variant; anything else throws TRANSLATION_SHAPE_INVALID with the same internal-fault reading, since every client-supplied path rejects a non-string entry before serving.

Tests: the delegation-equality checks move from a test of the abstract AbstractLayoutMutation into InsertElementTest, driven through the real scaffold path; the translatable reject rows that PropertyTypeTest owns are trimmed to one bare-string row in the two downstream test files, which keep the accept-the-map row that pins the shared predicate.
