+++ src/Core/Framework/ContentSystem/Binding/docs/symbols.md
@@ -1,0 +1,23 @@
+# Binding Symbols
+
+The symbol index for the binding system: each class, its role, and its path. The constraints that
+govern editing them stay in [../AGENTS.md](../AGENTS.md), and the conceptual model in
+[../README.md](../README.md).
+
+- **Value objects** (`Specification/`): `BindingSpecification` (`id`, `type`, `label`, `resolves`, `inputs`, `source`; `qualifiedId()` → `source:id`, `isDefault()`, `toSchema()`), `LoaderBinding` (`source`, `config`), `BindingInput` (`hasDefault (bool)`, `default (mixed)`, `required (bool)` — derived by the canonicalizer)
+- **resolvedBy loader branch**: `ResolvedByLoaderBranch` (module root; `@internal` enum, cases `Entity`/`EntityCollection`) — the closed tier-A classification: `fromReferenceFqcn(fqcn)` (`null` for neither), `loaderSource()` → the built-in loader's `SOURCE` constant, never a bare string literal, `fromLoaderSource(source)` (inverse), `matchesStoredValueShape(value)` (per-branch stored-value shape check)
+- **Default synthesizer**: `DefaultBindingSpecificationSynthesizer` (module root; `@internal`) — injected into `YamlBindingSpecificationLoader`. `synthesize(data, type, path): ?array` (`null` when the file declares no `resolvedBy` property). Enforces `MAX_ID_LENGTH = 255` on the minted id and the storage-key collision rules as hard load errors naming the file
+- **Registry**: `Registry/AbstractContentSystemBindingSpecificationRegistry` (decoration pattern; `all()` keyed by `source:id`, `byType()`, `get()`, `invalidate()`), `Registry/ContentSystemBindingSpecificationRegistry` (leaf aggregator over the tagged loaders), `Registry/CachedContentSystemBindingSpecificationRegistry` (`cache.system` decorator, cache key `content_system.binding_specifications`)
+- **Compiler Pass**: `Framework/DependencyInjection/CompilerPass/ContentSystemElementTypeCompilerPass` injects the discovered element-type directories into `YamlBindingSpecificationLoader` too. Any new throw there uses `DependencyInjectionException` — that namespace must not throw `ContentSystemException`
+- **Loaders**: `Loader/AbstractContentSystemBindingSpecificationLoader` (`load(): list<BindingSpecification>`), `Loader/YamlBindingSpecificationLoader` (filesystem), `Loader/DatabaseBindingSpecificationLoader` (active app bindings in prod, empty in dev), `Layout/Type/Loader/ElementTypeSourceDirectory` (`source`, `path`, `prefix`; shared with the type loader), `Loader/ResolvedBindingSpecificationDto` (`id`, `source`, `dto`; `toSpecification()`)
+- **Serializer**: `Serialization/BindingSpecificationSerializer` (`denormalize` array → DTO; `normalize` DTO → array)
+- **Canonicalizer**: `Serialization/BindingSpecificationCanonicalizer` — `canonicalize(dto, id, typeOverlay = [])` expands sugared `resolves` entries into canonical `{loader, config}` form and resolves the declared type **overlay-first, then registry**. Unknown type → `bindingSpecificationUnknownType` (400); non-deterministic sugar → `bindingSpecificationCanonicalizationFailed` (400); neither is a `CLIENT_DEFECT_CODE`
+- **Specification DTO**: `Specification/Dto/BindingSpecificationDto` (`#[WellFormedBindingSpecification]`; carries no id — the loader passes the `bindings:` map key to `toBindingSpecification(id, source)`), `Specification/Dto/BindingSpecificationDtoCollection` (`#[TypeConsistentBindingSpecification]`, plus a `typeOverlay` field the semantic validator reads)
+- **Validators**: `Validation/WellFormedBindingSpecification` (+ validator) — class-level on the DTO, structural shape only, no registry lookups; `Validation/TypeConsistentBindingSpecification` (+ validator) — **class-level on `BindingSpecificationDtoCollection`**
+- **Applicator**: `BindingApplicator` (final instance service, injects `Hydration/DataLoader/DataLoaderConfigSerializerProvider` and `Layout/Type/Registry/AbstractContentSystemElementTypeRegistry`) — `apply(element, specification, bindingSpecificationId): StoredElement` (overwrite) and `applyFillOnly()` (fill-only, same signature); both rebuild the element rather than mutating it. The registry is there for the `inputs` default seed: it resolves the target `PropertySpecification` so the default lands in the shape storage holds ([applying.md](applying.md))
+- **Mutation op**: `Mutation/Op/BindElement` (extends `Mutation/AbstractLayoutMutation`) — three `400`s: `bindingSpecificationNotFound`, `mutationTargetNotFound`, `bindingTypeMismatch`. `Mutation/Op/InsertElement` and `Mutation/Op/ReplaceElement` fill-apply the type's default via `AbstractLayoutMutation::resolveDefaultSpecification()`; `InsertElement` also applies a request's explicit `bindingSpecificationId` on top. Per-op contracts: [Mutation/docs/operations.md](../../Mutation/docs/operations.md)
+- **AttributionReconciler**: `AttributionReconciler` — the last of `Layout/LayoutWriteBoundary`'s three passes; re-derives every element's `attributedSpecifications` on every DAL write of the `layout` field, dropping a diverged or dangling attribution rather than erroring
+- **API endpoints**: the `bind()`/`insert()` actions of `Api/LayoutMutationController` and `Api/ContentLayoutMutationController`, both injecting the registry and the applicator; each `insert()` applies an optional request `bindingSpecificationId` onto the freshly scaffolded element. Paths, route names and the request contract: [Api/AGENTS.md](../../Api/AGENTS.md), [Api/docs/mutation-binding.md](../../Api/docs/mutation-binding.md)
+- **App integration**: `App/Aggregate/AppContentSystemBindingSpecification/AppContentSystemBindingSpecificationDefinition` (table `app_content_system_binding_specification`, `UNIQUE KEY` on `(app_id, name)`), `App/Lifecycle/Persister/ContentSystemBindingSpecificationPersister`, `App/Lifecycle/Handler/ContentSystemBindingSpecificationLifecycleHandler`, `App/Validation/ContentSystemBindingSpecificationAppValidator`, `App/Validation/Error/ContentSystemBindingSpecificationSchemaError`. What each does: [loading-and-apps.md](loading-and-apps.md)
+- **Migration**: `Migration/V6_7/Migration1782423128AddAppContentSystemBindingSpecificationTable` — creates `app_content_system_binding_specification` (`id`, `app_id` FK `ON DELETE CASCADE`, `name`, `schema` JSON, `hash`, timestamps)
+- **Diagnostics tie-ins**: `Resolution/CandidateOrigin::Stored`, `Diagnostics/ViolationCode::MismatchedReferenceType` — [Resolution](../../Resolution/AGENTS.md), [Diagnostics](../../Diagnostics/AGENTS.md)
