+++ src/Core/Framework/ContentSystem/ContentSystemException.php
@@ -399,13 +399,12 @@ class ContentSystemException extends HttpException
 
     /**
      * A translatable property holds one value per language as a language map, and serving collapses that map
-     * to the request language before any rendering step runs. A value that is not a map where the collapse
-     * runs is an internal fault rather than a client defect, and is deliberately absent from
+     * to the request language before any rendering step runs. A value that is not a map of strings where the
+     * collapse runs is an internal fault rather than a client defect, and is deliberately absent from
      * {@see self::CLIENT_DEFECT_CODES} — the same reading {@see invalidElementId()} and
      * {@see duplicateElementId()} state. Every client-supplied path rejects the wrong shape earlier, the
-     * strict write with a 400 and the draft routes with a reported violation, so a non-map value here means
-     * the write constraints were bypassed or a preparation listener introduced the shape after a conforming
-     * read.
+     * strict write with a 400 and the draft routes with a reported violation, so a wrong shape here means the
+     * write constraints were bypassed or a preparation listener introduced it after a conforming read.
      */
     public static function translationShapeInvalid(string $elementId, string $key, string $actualType): self
     {
