+++ src/Core/Framework/ContentSystem/Diagnostics/docs/violation-rules.md
@@ -1,0 +1,10 @@
+# Violation Rules
+
+What each per-violation check reports and when it fires. The invariants that govern editing the
+diagnostics core stay in [../AGENTS.md](../AGENTS.md); the scope and severity of every code are set
+by `ViolationCode`, indexed in [symbols.md](symbols.md).
+
+- `DanglingLanguage` (intrinsic, WARNING) reports one violation per language-map key on a translatable property that no `language` row carries, keyed on the property, naming the key in the message. It walks the map variant only — the other shapes are already `MismatchedPropertyType`. It gates nothing: key existence is deliberately not a write constraint, and `intrinsicErrors()` filters on scope AND severity (the band `OrphanedProvider` already occupied, so no filter changed). Re-read the language set on every `analyze()`; never memoize it on the instance, or a freshly created language reports as dangling.
+- `unfilledRequiredInputViolations` runs per resolution alongside `propertyBindingViolation` inside `bindingViolations`, and fires only on a required Reference resolved via `CandidateOrigin::Stored`. It encodes the stored `DataRequirement` config (`DataLoaderConfigSerializerProvider::encode`) and reads the source's config specification (`AbstractContentSystemDataLoaderMapResolver::resolve()->configSpecificationFor`), emitting one `UnfilledRequiredInput` per **required** `propertyReference` config key whose configured (string) value names an element property with no stored value (`!hasStoredValue($element, $configuredProperty)`: absent key and authored null both count, as does a translatable property whose language map lacks the anchor entry). Keyed on the input property when it is a declared primitive, else on the reference property (keying fallback for undeclared/non-primitive configured names). No loader source name appears — the rule keys off `ConfigKeyKind::PropertyReference` only. `encode()` is not wrapped in a client-defect catch: a Stored resolution proves the loader and (via the prior decode that built the config object) its serializer are registered, so a throw here is an internal fault that must surface.
+- `UnknownStyleOption` (intrinsic, error) is reported once per `style` key of an element that the style option registry's strict `all()` does not know, keyed on the option name. It reports; it never repairs — the read path still keeps an unknown option verbatim so an old layout renders, and the write still rejects it. Which surface shows the code differs: the diagnose route and the draft mutation routes carry it in their 200 diagnostics body (the persisted mutation routes commit before diagnosing, so the write rejects first); the preview route turns it into an `elementTypesInvalid` 400 carrying the message (not the code) via `DraftLayoutChecker`; the DAL write never produces it, because the `Layout/Codec/StoredTreeConstraints` descriptor rejects an unknown option during `encode()` and pre-empts the gate, ahead of the constraint-ordering rule in [../AGENTS.md](../AGENTS.md).
+- `storedRequirementViolation()` runs one `resolveType()` call per stored `DataRequirement` and reports exactly one outcome: a client-defect resolve failure is `InvalidConfig`; a resolve that succeeds but produces a type not assignable to the property's declared reference FQCN is `MismatchedReferenceType`; a resolve that succeeds and fits produces no violation (`Resolution/ElementResolver` reports it as a `Stored` resolution instead, never a `candidates` menu entry).
