+++ src/Core/Framework/ContentSystem/Event/Listener/README.md
@@ -5,14 +5,15 @@ Event-driven extension points for the content rendering lifecycle. A listener re
 ## Guides
 
 - [docs/custom-listeners.md](docs/custom-listeners.md) - The plugin-facing guide to writing a rendering-lifecycle listener.
+- [docs/listener-api.md](docs/listener-api.md) - The element members a listener edits, a worked listener, cache tags, and priority.
 
 ## Execution Order
 
-Core ships no listener on either event. `ContentPipeline::load()` (module root) calls its preparation and finishing steps directly, in this order:
+Core ships no listener on either event. `ContentPipeline::load()` (module root) calls its preparation and finishing steps directly, in this order — the rules that make the order load-bearing are owned by [../../docs/pipeline-steps.md](../../docs/pipeline-steps.md):
 
 After `ContentTreePreparationEvent` is dispatched:
-1. Language reduction — replaces each translatable property's language map with the value of the first entry along the request's language chain, in FULL mode only (`Layout/Scaffolding/StoredTreePreparer`, whose remaining steps follow); a map carrying no chain entry collapses to the null variant, which the mint skips
-2. Placeholder resolution — resolves `{{variable}}` placeholders from the specification on the stored tree, in FULL mode only, after the reduction: substitution touches string values only and never descends into a map, so a placeholder inside a translatable property resolves only once the map is collapsed to a plain string
+1. Language reduction — replaces each translatable property's language map with the value the request's language chain selects (`Layout/Scaffolding/StoredTreePreparer`, whose remaining steps follow)
+2. Placeholder resolution — resolves `{{variable}}` placeholders from the specification on the stored tree
 3. Virtual-root wrap: wraps the stored roots with a temporary container carrying the page-level placeholder values (`Layout/Scaffolding/VirtualRootWrapper`), after the placeholder resolution
 4. Partial prune — prunes the stored tree when `targetElementId` is specified (`Output/PartialRenderer`), after the virtual-root wrap; it ends the preparer's work, which hands back both the pruned tree and the forest as it stood before this step
 5. Duplicate-element-id check — rejects a repeated id (`CONTENT_SYSTEM__DUPLICATE_ELEMENT_ID`, 500), judging the pre-prune forest
@@ -31,4 +32,4 @@ So a `ContentTreePreparationEvent` listener always sees the raw loaded tree, and
 
 ## Priorities
 
-There are no reserved bands. Priority only orders extension listeners against each other on the same event. See [docs/custom-listeners.md](docs/custom-listeners.md) for the priority-band-inversion history.
+There are no reserved bands. Priority only orders extension listeners against each other on the same event. See [docs/listener-api.md](docs/listener-api.md) for what a plugin written against the old bands has to re-check.
