+++ src/Core/Framework/ContentSystem/Layout/Element/AGENTS.md
@@ -31,7 +31,7 @@ for the traversal contract and its slot-map constraint.
 ## Constraints
 
 - `StoredElement::property()` returns `null` for an absent key and never throws. An authored null is a PRESENT value whose variant is null, so it comes back as a `StoredValue` answering true to `isNull()` — absent and present-null are distinct states, and a bare null check conflates them
-- `StoredElement::properties()` never changes on a given instance — the map is private and `withProperties()` returns a new element. Its content is not only what an author typed: `Layout/LayoutDefaultSeeder` seeds a type's primitive defaults at the write boundary, and `Layout/Scaffolding/StoredTreePreparer` collapses each translatable property's language map to the request language and then substitutes placeholders into string values, both during preparation and in FULL mode only. Every value is wrapped in a `StoredValue`
+- `StoredElement::properties()` never changes on a given instance — the map is private and `withProperties()` returns a new element. Its content is not only what an author typed: `Layout/LayoutDefaultSeeder` seeds a type's primitive defaults at the write boundary, and `Layout/Scaffolding/StoredTreePreparer` rewrites property values during preparation ([../../docs/pipeline-steps.md](../../docs/pipeline-steps.md)). Every value is wrapped in a `StoredValue`
 - A property whose type declares `translatable: true` holds the MAP variant, not the string variant: one string per language keyed by language id in lowercase UUID hex, anchored on `Defaults::LANGUAGE_SYSTEM` ([README.md](README.md#translatable-properties)). Never write a bare string under such a key — the strict write rejects it (400) and reduction rejects it (`TRANSLATION_SHAPE_INVALID`, 500). `PropertyType::admits()` is the one predicate that answers whether a stored value fits its declared type, translatable or not
 - `Rendering/RenderedElementFactory` decides which keys the MINT produces: declared authored properties — every declared type except a single-FQCN reference, unions and bare `object` included — carrying the stored value, skipped when that value is the null variant, `dataRequirements[$key]` keys carrying the resolved loader value, the keys context was actually delivered under, and stored keys a parent's distribution config names — that last member excluding a declared reference property. Downstream is not closed: a `RenderedTreeFinalizationEvent` listener may hand back elements changed through `withProperty()` / `withProperties()`, and `ContentPipeline` carries that replacement forward rather than the tree it dispatched
 - A rendered element's own property map draws no distinction between a static, a loaded and a context-provided value; provenance is recorded beside the mint instead, in `Rendering/ElementMintResult` and the `LoweringResult` that collects them
