+++ src/Core/Framework/ContentSystem/Layout/Element/README.md
@@ -18,7 +18,7 @@ The storage-side element model, and the edit idiom for the rendered one. Element
 
 A property whose element type declares `translatable: true` stores one string per language rather than a bare value. Its `StoredValue` is the map variant, keyed by language id in lowercase UUID hex, and every entry value is a string variant. `Defaults::LANGUAGE_SYSTEM` is the anchor entry: it terminates every language chain a `SalesChannelContext` is built with, so an anchor entry resolves on every request while any other entry may not, and a required translatable property counts as filled only through it. The map carries at least one entry — an untranslated language simply has no key, and a property with no translation at all is absent from `properties()`.
 
-The map is the shape every stored-side reader sees, the admin round-trip included. Serving does not: language reduction, the first pass of `Layout/Scaffolding/StoredTreePreparer`, replaces each such map with the value of the first entry along the request's language chain, verbatim, before any other preparation step runs. A map carrying no chain entry reduces to the null variant, which the mint skips, so the property behaves as unset. Everything downstream of that pass — placeholder resolution, the lowering, the mint, every encoder — sees a plain string. Reduction runs in FULL mode only: SKELETON skips it and its mint carries no property value at all, so neither mode puts a language map on the wire.
+The map is the shape every stored-side reader sees, the admin round-trip included. Serving does not: language reduction, a pass of `Layout/Scaffolding/StoredTreePreparer`, replaces each such map with the value of the first entry along the request's language chain, verbatim. A map carrying no chain entry reduces to the null variant, which the mint skips, so the property behaves as unset. Neither rendering mode puts a language map on the wire. Where that pass sits among the preparation passes is owned by [../../docs/pipeline-steps.md](../../docs/pipeline-steps.md).
 
 ## Editing a Rendered Forest
