+++ src/Core/Framework/ContentSystem/Layout/README.md
@@ -6,7 +6,7 @@ Content layout tree structure and processing. Layouts are reusable templates con
 
 1. **Element Structure** (Element/) — the stored element model (`StoredElement`, `StoredValue`) plus `RenderedTreeEditor`, the whole-tree edit idiom for a rendered forest
 2. **DAL Definitions** (Entity/, Field/) — Database schema and custom field serializers
-3. **Scaffolding** (Scaffolding/) — The stored-tree preparer and the virtual-root wrapper it drives, plus the two records the preparer hands back: `TreePreparationResult` and, inside it, the `RenderScaffolding` carrying the wrap outcome to the finishing steps. Preparation opens with language reduction, the pass that collapses each translatable property's language map to the request language, in FULL mode only; the passes after it see plain strings. SKELETON skips reduction, and its mint carries no property value at all, so neither mode puts a language map on the wire
+3. **Scaffolding** (Scaffolding/) — The stored-tree preparer and the virtual-root wrapper it drives, plus the two records the preparer hands back: `TreePreparationResult` and, inside it, the `RenderScaffolding` carrying the wrap outcome to the finishing steps. Neither rendering mode puts a language map on the wire. The preparer's pass order is owned by [../docs/pipeline-steps.md](../docs/pipeline-steps.md)
 4. **Default Seeding** (`LayoutDefaultSeeder`) seeds element-type primitive defaults into the stored tree at the DAL write boundary, invoked from the `Field/` layout serializer's `normalize` hook
 
 ## Default Seeding
@@ -22,5 +22,5 @@ ContentLayoutEntity can contain multiple root elements. Each root is an independ
 - **[Element/](Element/README.md)** - Stored element model, the rendered-forest edit idiom, context and data requirement definitions
 - **Entity/** - DAL definitions (ContentLayoutDefinition)
 - **[Field/](Field/README.md)** - Custom DAL field types and serializers (infrastructure)
-- **Scaffolding/** - `StoredTreePreparer` (the one component that brings a stored forest into renderable shape: language reduction, then placeholder resolution, then the virtual-root wrap, then the partial prune, the first two in FULL mode only), `VirtualRootWrapper` (wraps the stored roots to carry the page-level placeholder values, recognises its own wrapper on the stored post-prune forest, and unwraps it again after the render step, as one of the finishing steps on the rendered forest), and the two records the preparer returns. `TreePreparationResult` carries the pruned tree, the pre-prune forest (which the pipeline's wiring validation judges, so a defect in a discarded subtree still fails the render) and the `RenderScaffolding`; `RenderScaffolding` is the immutable record the finishing steps read instead of re-deriving. Only `virtualRootSurvivedPrune` is read off the post-prune tree; `extractTargetId` is normalised from the `RenderingSpecification` before the prune, and the prune itself consumes it
+- **Scaffolding/** - `StoredTreePreparer` (the one component that brings a stored forest into renderable shape, in the pass order [../docs/pipeline-steps.md](../docs/pipeline-steps.md) owns), `VirtualRootWrapper` (wraps the stored roots to carry the page-level placeholder values, recognises its own wrapper on the stored post-prune forest, and unwraps it again after the render step, as one of the finishing steps on the rendered forest), and the two records the preparer returns. `TreePreparationResult` carries the pruned tree, the pre-prune forest (which the pipeline's wiring validation judges, so a defect in a discarded subtree still fails the render) and the `RenderScaffolding`; `RenderScaffolding` is the immutable record the finishing steps read instead of re-deriving. Only `virtualRootSurvivedPrune` is read off the post-prune tree; `extractTargetId` is normalised from the `RenderingSpecification` before the prune, and the prune itself consumes it
 - **[Type/](Type/README.md)** - Element type system: declarative type definitions, YAML loading, registry, app integration
