+++ src/Core/Framework/ContentSystem/Layout/Type/AGENTS.md
@@ -10,21 +10,13 @@ Type spec `properties` = schema for hydrated API output, NOT storage format
 - The storage format is published too, on the same endpoint: the `storageSchema` fold derived by `StoredSchemaResolver` (see below). `properties` is not it
 - Shared key links: type spec property key = dataRequirements key = acceptsContext key = the key `Rendering/RenderedElementFactory` writes the resolved value under
 
-## Source Code References
+## Navigation
 
-- **Registry**: `Registry/AbstractContentSystemElementTypeRegistry` (abstract, decoration pattern), `Registry/ContentSystemElementTypeRegistry` (stateless aggregator), `Registry/CachedContentSystemElementTypeRegistry` (`cache.system` pool decorator)
-- **Compiler Pass**: `Framework/DependencyInjection/CompilerPass/ContentSystemElementTypeCompilerPass` (discovers from core, bundles, plugins, apps; injects the directory set into both `YamlTypeLoader` and the binding system's `YamlBindingSpecificationLoader`)
-- **Loaders**: `Loader/AbstractContentSystemElementTypeLoader` (base contract), `Loader/YamlTypeLoader` (filesystem; also exposes `loadOverlayFromDirectory(directory, source, prefix): array<string, ContentSystemElementTypeSpecification>`, a registry-independent single-directory load keyed by resolved type name, see [Binding/docs/inline-bindings.md](../../Binding/docs/inline-bindings.md)), `Loader/DatabaseTypeLoader` (app types, prod only), `Loader/ElementTypeNameResolver` (path → name)
-- **Serializer**: `Serialization/ElementTypeSpecificationSerializer` (YAML ↔ DTO)
-- **API Endpoint**: `Api/Controller/InfoController::getContentSystemElementTypes()` (`GET /api/_info/content-system-element-types.json`)
-- **Storage schema**: `StoredSchemaResolver::resolve(ContentSystemElementTypeSpecification): array<string, array{kind, type, required, default?, translatable?}>` — the per-type `storageSchema` fold on that endpoint, keyed by STORED key. `translatable` is emitted, always as `true`, only on a `property` entry whose declared property is translatable, and omitted otherwise the way `default` is; `default` beside it stays the declared scalar rather than the anchor map storage seeds. See [docs/introspection.md#storageschema](docs/introspection.md#storageschema) for the three kinds, their precedence, and the typing rules
-- **App Integration**: `App/Lifecycle/Persister/ContentSystemElementTypePersister`, `App/Validation/ContentSystemElementTypeAppValidator`, `App/Lifecycle/Handler/ContentSystemElementTypeLifecycleHandler` (persists app types on install/update)
-- **Collision Detection**: `Validation/ElementTypeCollisionDetector` (validates proposed names against registry + inactive app types)
-- **Type Map Bridge**: `Schema/ContentSystemDataLoaderMap` — connects FQCNs to loader sources
-- **Primitive defaults**: `PrimitiveDefaultProvider::forType(AbstractContentSystemElementTypeRegistry $registry, string $type): array<string, string|int|float|bool|array<string, string|int|float|bool>>`. The single per-type rule, returning the non-null `storedDefault()` of every primitive (`isPrimitive()`) property keyed by property key — the array member of the return type is the language map a translatable default seeds as. The one definition of "a type's primitive defaults": `Mutation/AbstractLayoutMutation::primitiveDefaults()` (scaffold + replace seeding) and `Layout/LayoutDefaultSeeder::defaultsFor()` (write-boundary seeding) both DELEGATE here and keep no copy, so the two agree by construction. The caller guarantees the type is registered
-- **Stored shape rule**: `Specification/PropertyType::storedDefault(): string|int|float|bool|array|null` is the one statement of what shape a declared default takes in storage — the bare scalar for every property, `{Defaults::LANGUAGE_SYSTEM: default}` for a translatable one. It wraps only a non-null default; a declaration with no default returns `null` and seeds nothing, which is `PrimitiveDefaultProvider`'s skip. It lives on the type beside the flag it reads. A binding specification's `inputs` default is a different value declared elsewhere, so `Binding/BindingApplicator` applies the same rule rather than calling this producer ([Binding/AGENTS.md](../../Binding/AGENTS.md))
-- **Conformance predicate**: `Specification/PropertyType::admits(StoredValue $value): bool` is the single answer to whether a stored value fits its declared type, total over every `StoredValue` variant including null. Translatable: a non-empty map whose every entry is a string variant, and nothing else — not a bare string, not the empty map, not the null variant. Otherwise: the primitive match table, where a union carrying a non-primitive, a bare `object` and an FQCN constrain nothing, and the null variant is admitted throughout. `Specification/PropertyType::translatable(): bool` exposes the flag, which `toSchema()` otherwise hides. `Layout/Codec/PropertyTypeConformanceValidator`, `Diagnostics/LayoutDiagnostics` and `Mutation/Op/ReplaceElement` all call `admits()` and none keeps a match table — do not reintroduce one
-- **Specification accessors**: `Specification/ContentSystemElementTypeSpecification` exposes `name(): string` (the type's unique identifier — the key consumers match/look up by), `source(): string` (the source-label prefix, e.g. `core` / `plugin:Name`, default `''`), `properties(): array<string, PropertySpecification>` (the declared property map the `Resolution/` kernel — `Resolution/ElementResolver`, `Resolution/AvailableContextResolver` — iterates to resolve each property, reading each `PropertySpecification` and its `PropertyType::isPrimitive()`), `slots(): list<SlotSpecification>` (the declared slots; each `SlotSpecification` exposes `name(): string`, consumed by `Mutation/Op/ReplaceElement` to decide which slot children a new type can keep), and `toSchema(): array` (serializes the spec to the `ElementTypeSchema` wire shape served by `GET /api/_info/content-system-element-types.json`)
+- Symbol index: classes, roles, paths — [docs/symbols.md](docs/symbols.md)
+- The five stages a type declaration passes through — [docs/architecture.md](docs/architecture.md)
+- The `/api/_info/` type endpoint and the `storageSchema` fold — [docs/introspection.md](docs/introspection.md)
+- Why the type spec is the output schema, not the storage format — [docs/output-schema.md](docs/output-schema.md)
+- Authoring a custom element type — [docs/custom-types.md](docs/custom-types.md)
 
 ## Constraints
