+++ src/Core/Framework/ContentSystem/docs/layout-write-gates.md
@@ -7,4 +7,4 @@ stays with the directory that owns it.
 - **Resolvability gate** (DAL `PreWriteValidationEvent`): two subscribers enforce served-implies-resolvable — `Validation/ContentLayoutWriteValidator`, the single ordered `content_layout` write gate, and `Validation/ContentLayoutAssignmentWriteValidator`, a tree-blind type-match on the bound layout's immutable `root_source`. `Validation/LayoutGate` holds the two predicates and the `SKIP_VALIDATION_STATE` bypass. Gate ordering, the edit-path membership re-check, every error code, and the served-implies-resolvable stationarity caveat are owned by [Validation/AGENTS.md](../Validation/AGENTS.md) / [Validation/README.md](../Validation/README.md)
 - **Write-boundary default seeding**: `Layout/LayoutWriteBoundary` is the one admission point every layout write passes through, reached from the DAL write chokepoint for the `layout` field (`Layout/Field/StoredElementListFieldSerializer::normalize`). `normalize` runs before `PreWriteValidationEvent` and memoizes the admitted tree on the write `Context`, so the gate judges a seeded tree and the write validator stays a pure gate that never rewrites. Migrations write raw SQL via `Connection` and bypass this seam. What the three passes do — `Layout/LayoutDefaultSeeder` and its shared per-type rule included — is owned by [Layout/README.md](../Layout/README.md) and [Binding/docs/write-boundary.md](../Binding/docs/write-boundary.md)
 - **Delete protection**: a `content_layout` bound to any source cannot be deleted — `RestrictDelete` on the five `content_layout` OneToMany assignment associations (`Layout/Entity/ContentLayoutDefinition` for product/category/landing-page, `Storefront/.../Extension/ContentLayoutExtension` for header/footer) makes the DAL throw `RestrictDeleteViolationException` before any SQL, backed by `ON DELETE RESTRICT` on each `content_layout_id` foreign key. This is DAL-core machinery, not a content-system gate, so `SKIP_VALIDATION_STATE` does not bypass it. Sibling foreign keys (`sales_channel_id`, `domain_id`) stay `ON DELETE CASCADE` — deleting the referenced channel/domain still removes the assignment while the layout survives
-- Primitive property satisfaction is strict: `Diagnostics/LayoutDiagnostics` reports a required primitive as `UnresolvedRequired` iff the element has no stored value for it (`LayoutDiagnostics::hasStoredValue()`); the type default is not consulted, because serving renders the stored `properties` verbatim. For a TRANSLATABLE property the anchor rule supersedes present-and-not-null: satisfaction needs the property's language map to carry `Defaults::LANGUAGE_SYSTEM` with a string value, so a present, non-null map no longer satisfies on its own, and the same rule governs `UnfilledRequiredInput` because both violations read the one `hasStoredValue()`. A default reaches storage only through creation-time seeding (`Mutation` scaffold/replace) and the write-boundary `Layout/LayoutDefaultSeeder`, so a writer that bypasses the DAL (raw SQL, migration) must seed a required defaulted primitive itself or the next resolvability re-check rejects the row, and the unseeded draft path (`Api/ContentDiagnoseController`, preview) deliberately over-reports it. See [Diagnostics/AGENTS.md](../Diagnostics/AGENTS.md)
+- Primitive property satisfaction is strict: `Diagnostics/LayoutDiagnostics` reports a required primitive as `UnresolvedRequired` iff the element has no stored value for it (`LayoutDiagnostics::hasStoredValue()`); the type default is not consulted, because serving renders the stored `properties` verbatim. A translatable property answers that same question under the anchor rule instead. A default reaches storage only through creation-time seeding (`Mutation` scaffold/replace) and the write-boundary `Layout/LayoutDefaultSeeder`, so a writer that bypasses the DAL (raw SQL, migration) must seed a required defaulted primitive itself or the next resolvability re-check rejects the row, and the unseeded draft path (`Api/ContentDiagnoseController`, preview) deliberately over-reports it. See [Diagnostics/AGENTS.md](../Diagnostics/AGENTS.md)
