+++ src/Core/Framework/ContentSystem/docs/pipeline-steps.md
@@ -5,6 +5,6 @@ load-bearing. The symbols are indexed in [../AGENTS.md](../AGENTS.md); the diagr
 [data-flow.md](data-flow.md).
 
 - `ContentPipeline::load()`: receives a loaded `RenderableLayout` (stored elements) → `ContentTreePreparationEvent` → stored preparation (`Layout/Scaffolding/StoredTreePreparer::prepare()`, which takes the `SalesChannelContext`: language reduction, then placeholder resolution — both inside one FULL-mode guard, in that order — then the virtual-root wrap via `Layout/Scaffolding/VirtualRootWrapper`, then the partial prune via `Output/PartialRenderer`, all still stored, returning a `Layout/Scaffolding/TreePreparationResult`) → duplicate-element-id rejection on the result's pre-prune forest → `Rendering/WiringPlanner::plan()` (wiring validation on that same pre-prune forest, then redistribute derivation on the result's pruned tree) → the render step (`Rendering/ElementLowering::lower()`: forest-wide data resolution, the once-per-render root-ambient resolution of the specification's page-level data requirements against the virtual-root wrapper the pipeline hands it from the pre-prune forest, context-delivery resolution, rendered-tree minting; FULL does all four, SKELETON mints structure only) → the finishing steps, on the rendered forest (`VirtualRootWrapper::unwrap()`, then `PartialRenderer::extractTarget()`, both driven by the `RenderScaffolding`) → `RenderedTreeFinalizationEvent` over the rendered forest, in both modes → duplicate-element-id rejection again, over the forest the event handed back, because a finalization listener may replace it and the pre-lowering check ran too early to see that. Both checks throw `DUPLICATE_ELEMENT_ID` (500, not a client defect) and run in either mode. It returns an `Output/RenderResult` (finished rendered forest, layout reference, optional `Output/Index/ResolvedValueIndex`), and takes both format answers from the route: the `RenderingMode` and whether the format collects a value index
-- Language reduction is the first preparation pass, and its position is load-bearing. It replaces each translatable property's language map with the value of the first entry along `SalesChannelContext::getLanguageIdChain()`, verbatim; a key outside the chain is never selected, so a dangling language id cannot reach serving, and a map carrying no chain entry collapses to the null variant, which the mint skips. Placeholder resolution runs second because substitution touches string values and never descends into a map. In FULL mode, everything after it reads plain strings: the lowering, the mint, every encoder, the value index and the Twig components. SKELETON reaches the same end by a different route — reduction is skipped, so `Rendering/WiringPlanner` (which runs in both modes) receives the unreduced maps and is indifferent to them because it reads no property value, and the SKELETON mint strips properties entirely. Neither mode puts a language map on the wire. A non-map value on a translatable property at reduction throws `TRANSLATION_SHAPE_INVALID` (500, deliberately absent from `CLIENT_DEFECT_CODES` — the same internal-fault reading as `DUPLICATE_ELEMENT_ID`): every client-supplied path rejects that shape earlier, so reaching it means the write constraints were bypassed or a `ContentTreePreparationEvent` listener introduced it after a conforming read
+- Language reduction is the first preparation pass, and its position is load-bearing. It replaces each translatable property's language map with the value of the first entry along `SalesChannelContext::getLanguageIdChain()`, verbatim; a key outside the chain is never selected, so a dangling language id cannot reach serving, and a map carrying no chain entry collapses to the null variant, which the mint skips. Placeholder resolution runs second because substitution touches string values and never descends into a map. In FULL mode, everything after it reads plain strings: the lowering, the mint, every encoder, the value index and the Twig components. SKELETON reaches the same end by a different route — reduction is skipped, so `Rendering/WiringPlanner` (which runs in both modes) receives the unreduced maps and is indifferent to them because it reads no property value, and the SKELETON mint strips properties entirely. Neither mode puts a language map on the wire. A non-map value on a translatable property at reduction, or a selected entry that is not a string, throws `TRANSLATION_SHAPE_INVALID` (500, deliberately absent from `CLIENT_DEFECT_CODES` — the same internal-fault reading as `DUPLICATE_ELEMENT_ID`): every client-supplied path rejects that shape earlier, so reaching it means the write constraints were bypassed or a `ContentTreePreparationEvent` listener introduced it after a conforming read
 - Wiring validation and redistribute derivation are the two halves of `Rendering/WiringPlanner::plan()`, deliberately split: validation judges the pre-prune forest (a defect in a subtree a partial render discards still fails the render), derivation runs on the surviving tree and throws nothing
 - Specification resolution and layout-entity loading happen in `ContentRoute`, NOT in `ContentPipeline`
